week3/community-contributions/Mksong/requirements.txt
# Core ML libraries
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1

# HuggingFace ecosystem
transformers>=4.48.3
datasets>=3.2.0
diffusers
accelerate>=1.3.0
bitsandbytes>=0.46.0
huggingface_hub

# Audio processing
soundfile
openai-whisper

# Image processing
Pillow
# Utilities
python-dotenv
requests


